[RELEASE] Custom jobs for FiveM.
Custom ESX Jobs is a resource created by KEVIN which combine all jobs in one working script. By modifying ESX and expanding es_extended, it is possible to work with permissions and adjust everything in real time. Jobs can be edited or added while the server is running (deleting a job must be done manually).